embeddings-retrieval/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
filelock==3.19.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
packaging==25.0
pillow==11.3.0
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.7.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.67.1
transformers==4.44.2
typing_extensions==4.15.0
urllib3==2.5.0
faiss-cpu==1.8.0
